server/requirements.txt
# Core dependencies
fastapi
uvicorn
python-multipart
pillow
pdfplumber
pytesseract
python-dotenv

# Haystack and Integrations
haystack-ai
google-ai-haystack
sentence-transformers

google-generativeai

easyocr